+++ routes/api.php
@@ -29,20 +29,20 @@ Route::name('app')->namespace('App\Http\Controllers')->group(function() {
 
     Route::prefix('/users')->group(function() {
         Route::get('/', [UserController::class, 'index']);
-        Route::get('/{id}', [UserController::class, 'show']);
-        Route::post('/', [UserController::class, 'create']);
-        Route::put('/{id}', [UserController::class, 'update']);
-        Route::delete('/{id}', [UserController::class, 'destroy']);
+        Route::get('/{id}', [UserController::class, 'show'])->middleware('jwt.auth');
+        Route::post('/', [UserController::class, 'store']);
+        Route::put('/{id}', [UserController::class, 'update'])->middleware('jwt.auth');
+        Route::delete('/{id}', [UserController::class, 'destroy'])->middleware('jwt.auth');
         Route::post('/login', [JwtController::class, 'login']);
-        Route::get('/refresh-session', [JwtController::class, 'refresh']);
-        Route::get('/logout', [UserController::class, 'logout']);
+        Route::get('/refresh-session', [JwtController::class, 'refresh'])->middleware('jwt.auth');
+        Route::get('/logout', [UserController::class, 'logout'])->middleware('jwt.auth');
     });
 
 
     Route::prefix('/expense-category')->group(function() {
         Route::get('/', [ExpenseCategoryController::class, 'index']);
         Route::get('/{id}', [ExpenseCategoryController::class, 'show']);
-        Route::post('/', [ExpenseCategoryController::class, 'create']);
+        Route::post('/', [ExpenseCategoryController::class, 'store']);
         Route::put('/{id}', [ExpenseCategoryController::class, 'update']);
         Route::delete('/{id}', [ExpenseCategoryController::class, 'destroy']);
     });
@@ -51,7 +51,7 @@ Route::name('app')->namespace('App\Http\Controllers')->group(function() {
     Route::prefix('income-category')->group(function() {
         Route::get('/', [IncomeCategoryController::class, 'index']);
         Route::get('/{id}', [IncomeCategoryController::class, 'show']);
-        Route::post('/', [IncomeCategoryController::class, 'create']);
+        Route::post('/', [IncomeCategoryController::class, 'store']);
         Route::put('/{id}', [IncomeCategoryController::class, 'update']);
         Route::delete('/{id}', [IncomeCategoryController::class, 'destroy']);
     });
@@ -60,7 +60,7 @@ Route::name('app')->namespace('App\Http\Controllers')->group(function() {
     Route::prefix('/expenses')->group(function() {
         Route::get('/', [ExpensesController::class, 'index']);
         Route::get('/{id}', [ExpensesController::class, 'show']);
-        Route::post('/', [ExpensesController::class, 'create']);
+        Route::post('/', [ExpensesController::class, 'store']);
         Route::put('/{id}', [ExpensesController::class, 'update']);
         Route::delete('/{id}', [ExpensesController::class, 'destroy']);
     });
@@ -69,7 +69,7 @@ Route::name('app')->namespace('App\Http\Controllers')->group(function() {
     Route::prefix('/incomes')->group(function() {
         Route::get('/', [IncomesController::class, 'index']);
         Route::get('/{id}', [IncomesController::class, 'show']);
-        Route::post('/', [IncomesController::class, 'create']);
+        Route::post('/', [IncomesController::class, 'store']);
         Route::put('/{id}', [IncomesController::class, 'update']);
         Route::delete('/{id}', [IncomesController::class, 'destroy']);
     });
